+++ spec/rails_helper.rb
@@ -33,6 +33,8 @@ RSpec.configure do |config|
   config.infer_base_class_for_anonymous_controllers = false
   config.infer_spec_type_from_file_location!
 
+  config.include Devise::Test::IntegrationHelpers, type: :request
+
   # DatabaseCleaner set up
   config.use_transactional_fixtures = false
   config.before(:suite) do